requirements.txt
gradio==6.13.0
numpy==1.26.4
opencv-python==4.10.0.84
Pillow==12.2.0
realesrgan==0.3.0
basicsr==1.4.2
torch==2.0.1
torchvision==0.15.2
